requirements.txt
telethon==1.34.0
aiohttp==3.9.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
